requirements.txt
click==8.1.3
colorama==0.4.6
Flask==2.2.5
greenlet==2.0.2
gunicorn==20.1.0
importlib-metadata==6.6.0
itsdangerous==2.1.2
Jinja2==3.1.2
logging==0.4.9.6
MarkupSafe==2.1.2
psycopg2==2.9.6
pydantic==1.10.8
python-decouple==3.8
python-dotenv==0.21.1
SQLAlchemy==2.0.15
typing_extensions==4.6.2
Werkzeug==2.2.3
zipp==3.15.0
